Script to stop minikube when bootstrap node reboots
Can be used to simulate the scenario where assisted-service is
running on bootstrap node and becomes unavailable when the node
reboots to join the cluster.

In the simulated environment, the "bootstrap" assisted-service
runs on minikube as deployed by assisted-test-infra.

The script also downloads the cluster's kubeconfig before
minikube is stopped.